pyproject.toml
[project]
name = "qbw-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.5",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.1",
    "fastembed>=0.7.4",
    "ipykernel>=7.1.0",
    "langchain>=1.1.2",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.1",
    "langchain-openai>=1.1.3",
    "pymupdf>=1.26.6",
    "pypdf>=6.4.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "tqdm>=4.67.1",
    "watchdog>=6.0.0",
]
